chore: move claim reward logic in separate provider
